CNC-LandingPage3-2_white1/chatbot-server/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
mysql-connector-python==8.2.0

# Data processing
numpy==1.26.3
pandas==2.1.4

# AI/ML
openai==1.40.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.0
